+++ packages/components/src/components/button/button.tsx
@@ -19,7 +19,14 @@ import {
   Method,
 } from '@stencil/core';
 import classNames from 'classnames';
-import { hasShadowDom } from '../../utils/utils';
+import { hasShadowDom, ScaleIcon, isScaleIcon } from '../../utils/utils';
+
+const DEFAULT_ICON_SIZE = 24;
+
+const buttonIconSizeMap = {
+  small: 16,
+  // large: 24,
+};
 
 @Component({
   tag: 'scale-button',
@@ -47,8 +54,10 @@ export class Button {
   @Prop() target?: string = '_self';
   /** (optional) Injected CSS styles */
   @Prop() styles?: string;
-  /** (optional) name of a file to be downloaded */
+  /** (optional) Name of a file to be downloaded */
   @Prop() download?: string;
+  /** (optional) Set `tabindex` in the inner button or link element */
+  @Prop() innerTabindex?: number;
 
   private focusableElement: HTMLElement;
 
@@ -95,6 +104,10 @@ export class Button {
     this.setIconPositionProp();
   }
 
+  componentDidLoad() {
+    this.setChildrenIconSize();
+  }
+
   /**
    * Detect whether the last node is an element (not text).
    * If so, it's probably an icon, so we set `iconPosition` to `after`.
@@ -104,14 +117,25 @@ export class Button {
       // ignore empty text nodes, which are probably due to formatting
       return !(node.nodeType === 3 && node.nodeValue.trim() === '');
     });
-    if (
-      !this.iconOnly &&
-      nodes &&
-      nodes.length &&
-      nodes[nodes.length - 1] &&
-      nodes[nodes.length - 1].nodeName.substr(0, 10) === 'SCALE-ICON'
-    ) {
+    const lastNode = nodes.length > 1 ? nodes[nodes.length - 1] : null;
+    if (!this.iconOnly && lastNode && isScaleIcon(lastNode)) {
       this.iconPosition = 'after';
+    }
+  }
+
+  /**
+   * Set any children icon's size according the button size.
+   */
+  setChildrenIconSize() {
+    if (this.size != null && buttonIconSizeMap[this.size] != null) {
+      const icons: ScaleIcon[] = Array.from(this.hostElement.children).filter(
+        isScaleIcon
+      );
+      icons.forEach((icon) => {
+        if (icon.size === DEFAULT_ICON_SIZE) {
+          icon.size = buttonIconSizeMap[this.size];
+        }
+      });
     }
   }
 
@@ -137,6 +161,7 @@ export class Button {
             target={this.target}
             rel={this.target === '_blank' ? 'noopener noreferrer' : undefined}
             part={basePart}
+            tabIndex={this.innerTabindex}
           >
             <slot />
           </a>
@@ -148,6 +173,7 @@ export class Button {
             disabled={this.disabled}
             type={this.type}
             part={basePart}
+            tabIndex={this.innerTabindex}
           >
             <slot />
           </button>