+++ packages/components/src/components/menu-flyout-list/menu-flyout-list.tsx
@@ -23,7 +23,7 @@ import {
   Watch,
 } from '@stencil/core';
 import classNames from 'classnames';
-import { emitEvent, isClickOutside } from '../../utils/utils';
+import { emitEvent } from '../../utils/utils';
 
 const PAD = 10;
 const ITEM_ROLES = ['menuitem', 'menuitemcheckbox', 'menuitemradio'];
@@ -140,7 +140,11 @@ export class MenuFlyoutList {
     if (!this.active) {
       return;
     }
-    event.preventDefault();
+
+    if (!this.hostElement.querySelector('app-navigation-user-menu')) {
+      event.preventDefault();
+    }
+
     if ('ArrowDown' === event.key) {
       this.shiftItemsFocus();
       return;
@@ -180,13 +184,6 @@ export class MenuFlyoutList {
     if (item != null) {
       event.stopImmediatePropagation();
       item.triggerEvent('click', null, this.closeOnSelect);
-    }
-  }
-
-  @Listen('click', { target: 'document', capture: true })
-  handleDocumentClick(event: MouseEvent) {
-    if (isClickOutside(event, this.hostElement) && this.opened) {
-      this.close(true);
     }
   }