+++ packages/storybook-vue/.storybook/main.js
@@ -4,6 +4,7 @@ module.exports = {
     "@storybook/addon-links",
     "@storybook/addon-docs",
     "@storybook/addon-controls",
+    '@storybook/addon-viewport',
     "./usage-addon/register.js",
     "./language-addon/register.js",
     "./sidebar-links-addon/register.js",