+++ packages/storybook-vue/.storybook/usage-addon/usage.js
@@ -47,6 +47,10 @@ import menuFlyout_en from 'raw-loader!../../stories/3_components/flyout-menu/men
 import menuFlyout_de from 'raw-loader!../../stories/3_components/flyout-menu/menu-flyout_de.md';
 import modal_en from 'raw-loader!../../stories/3_components/modal/modal.md';
 import modal_de from 'raw-loader!../../stories/3_components/modal/modal_de.md';
+import notificationBadge_de from 'raw-loader!../../stories/3_components/notification-badge/notification-badge_de.md';
+import notificationBadge_en from 'raw-loader!../../stories/3_components/notification-badge/notification-badge.md';
+import notificationBanner_de from 'raw-loader!../../stories/3_components/notification-banner/notification-banner_de.md';
+import notificationBanner_en from 'raw-loader!../../stories/3_components/notification-banner/notification-banner.md';
 import notificationMessage_de from 'raw-loader!../../stories/3_components/notification-message/notification-message_de.md';
 import notificationMessage_en from 'raw-loader!../../stories/3_components/notification-message/notification-message.md';
 import pagination_en from 'raw-loader!../../stories/3_components/pagination/pagination.md';
@@ -83,6 +87,8 @@ import callout_en from 'raw-loader!../../stories/3_components/callout/callout.md
 import callout_de from 'raw-loader!../../stories/3_components/callout/callout_de.md';
 import toggleGroup_de from 'raw-loader!../../stories/3_components/toggle-group/toggle-group_de.md';
 import toggleGroup_en from 'raw-loader!../../stories/3_components/toggle-group/toggle-group.md';
+import tooltip_en from 'raw-loader!../../stories/3_components/tooltip/tooltip.md';
+import tooltip_de from 'raw-loader!../../stories/3_components/tooltip/tooltip_de.md';
 
 const NOT_A_COMPONENT_MD = '`Browse to any component to see usage.`';
 const COMPONENT_NOT_MAPPED_MD =
@@ -143,6 +149,10 @@ const Usage = (props) => {
     'flyout-menu_de': menuFlyout_de,
     modal_en,
     modal_de,
+    'notification-badge_en': notificationBadge_en,
+    'notification-badge_de': notificationBadge_de,
+    'notification-banner_en': notificationBanner_en,
+    'notification-banner_de': notificationBanner_de,
     'notification-message_en': notificationMessage_en,
     'notification-message_de': notificationMessage_de,
     pagination_en,
@@ -177,6 +187,8 @@ const Usage = (props) => {
     footer_de,
     'toggle-group_en': toggleGroup_en,
     'toggle-group_de': toggleGroup_de,
+    tooltip_en,
+    tooltip_de,
   };
 
   // Select the most appropriate markdown text based on if this is a component story and